mar y paises calculados por separado

+++ tdi.py
@@ -9,30 +9,55 @@ from perlinNice import *
 
 testsFolder = "tests"
 
-def generateMap(size, name):
+def generateMap(size, name, perlinContry, perlinSee, threshold, thresholdGrowth, seeLevel, medianSize, seeMedianSize):
   name = os.path.join(testsFolder,name)
 
   #arr = perlinNoiseFreq(size, [(2,2),(5,1)])
-  arr = perlinNoise(size, [(2,1),(6, 0.5), (10, 0.2)])
-  arr = np.array(256*arr, dtype = 'uint8') #pasamos de [0,1) a [0,255]
-  img.imsave(f'{name}_0perlinNoise.png', arr)
+  perlin = perlinNoise(size, perlinContry)
+  perlin = np.array(256*perlin, dtype = 'uint8') #pasamos de [0,1) a [0,255]
+  img.imsave(f'{name}_0perlinNoise.png', perlin)
 
-  # _, arr = cv2.threshold(arr, 100, 256, cv2.THRESH_TOZERO)
-  # # arr, treshold, maxvalue, type
-  # img.imsave(f'{name}_1umbralizado.png', arr)
+  arr = regionGrower(perlin,10,threshold, thresholdGrowth)
+  img.imsave(f'{name}_1regionGrower.png', arr)
 
-  arr = regionGrower(arr,10,threshold=1, thresholdGrowth=1.1)
-  img.imsave(f'{name}_2regionGrower.png', arr)
-
-  medianSize = size//100
+  medianSize = size//medianSize
   medianSize = medianSize+medianSize%2+1
   arr = cv2.medianBlur(arr, medianSize)
-  img.imsave(f'{name}_3filtroMediana.png', arr)
+  img.imsave(f'{name}_2filtroMediana.png', arr)
 
+
+  # see = perlinNoise(size, perlinSee)
+  # see = np.array(256*see, dtype = 'uint8') #pasamos de [0,1) a [0,255]
+  # img.imsave(f'{name}_5see.png', see)
+
+  
+  seeMedianSize = size//seeMedianSize
+  seeMedianSize = seeMedianSize+seeMedianSize%2+1
+
+  see = cv2.GaussianBlur(perlin, (seeMedianSize,seeMedianSize), cv2.BORDER_DEFAULT)
+  img.imsave(f'{name}_5seeGaussianBlur.png', see)
+
+  _, see = cv2.threshold(see, seeLevel, 256, cv2.THRESH_BINARY_INV)
+  img.imsave(f'{name}_6seeUmbralizado.png', see)
+
+  see = cv2.medianBlur(see, seeMedianSize)
+  img.imsave(f'{name}_7seeMediana.png', see)
+
+
+  arr = arr*see
+  img.imsave(f'{name}_8mareNostrum.png', arr)
   print(name,"terminado")
 
 if __name__ == "__main__":
-  shutil.rmtree(testsFolder)
+  if os.path.exists(testsFolder):
+    shutil.rmtree(testsFolder)
   os.makedirs(testsFolder)
   for i in range(10):
-    generateMap(200, str(i))
+    generateMap(200, str(i),
+      perlinContry=[(10, 1), (20, 0.2), (100, 0.2)],
+      perlinSee=[(2, 2), (5,1),(10, 1), (30, 0.05)],
+      threshold=1,
+      thresholdGrowth=1.1,
+      seeLevel=130,
+      medianSize=100,
+      seeMedianSize=10)
